requirements.txt
asgiref==3.3.4
click==8.0.1
colorama==0.4.4
Django==3.2.2
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
djangorestframework==3.12.4
joblib==1.0.1
nltk==3.6.2
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
tqdm==4.61.0
